Breakpoint added to check the root node.
The breakpoint is added to check the data of the root node after the rotation has been performed if the nodes are unbalanced. You can change the insert order to try out various permutations of the trees and use RecursiveInsert function to insert the values and then call RecursiveInsert function from the main() function.